Restructure embeddings
